+++ common/src/main/java/com/faboslav/friendsandfoes/client/render/entity/animation/animator/ModelPartAnimationType.java
@@ -1,0 +1,11 @@
+package com.faboslav.friendsandfoes.client.render.entity.animation.animator;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+
+@Environment(EnvType.CLIENT)
+public enum ModelPartAnimationType
+{
+	POSITION,
+	ROTATION
+}