+++ common/src/main/java/com/faboslav/friendsandfoes/util/plugin/MixinExtrasPlugin.java
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
 import java.util.List;
 import java.util.Set;
 
-public class MixinExtrasPlugin implements IMixinConfigPlugin
+public final class MixinExtrasPlugin implements IMixinConfigPlugin
 {
 	public void onLoad(String mixinPackage) {
 		MixinExtrasBootstrap.init();