+++ fabric/src/main/java/com/faboslav/friendsandfoes/platform/fabric/BiomeModificationsImpl.java
@@ -1,11 +1,16 @@
 package com.faboslav.friendsandfoes.platform.fabric;
 
+import com.faboslav.friendsandfoes.FriendsAndFoes;
 import net.fabricmc.fabric.api.biome.v1.BiomeModifications;
+import net.fabricmc.fabric.api.biome.v1.ModificationPhase;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.SpawnGroup;
+import net.minecraft.registry.RegistryKey;
+import net.minecraft.registry.RegistryKeys;
 import net.minecraft.registry.tag.BiomeTags;
 import net.minecraft.registry.tag.TagKey;
 import net.minecraft.world.biome.Biome;
+import net.minecraft.world.gen.GenerationStep;
 
 public final class BiomeModificationsImpl
 {
@@ -18,5 +23,20 @@ public final class BiomeModificationsImpl
 		int maxGroupSize
 	) {
 		BiomeModifications.addSpawn(biomeSelector -> biomeSelector.hasTag(tag) && biomeSelector.hasTag(BiomeTags.IS_OVERWORLD), spawnGroup, entityType, weight, minGroupSize, maxGroupSize);
+	}
+
+	public static void addButtercupFeature() {
+		TagKey<Biome> flowerForestTag = TagKey.of(RegistryKeys.BIOME, FriendsAndFoes.makeID("has_buttercup_patch"));
+		BiomeModifications.create(FriendsAndFoes.makeID("add_buttercup_patch"))
+			.add(ModificationPhase.ADDITIONS,
+				(context) -> context.hasTag(flowerForestTag),
+				context -> context.getGenerationSettings().addFeature(
+					GenerationStep.Feature.VEGETAL_DECORATION,
+					RegistryKey.of(
+						RegistryKeys.PLACED_FEATURE,
+						FriendsAndFoes.makeID("buttercup_patch")
+					)
+				)
+			);
 	}
 }